requirements.txt
h5py==3.9.0
layers==0.1.5
matplotlib==3.7.3
mypath==0.1
numpy==1.24.3
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.0.1
Pillow==10.4.0
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.3.1
scipy==1.14.1
setuptools==69.5.1
six==1.16.0
skimage==0.0
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
torch==2.2.2+cu121
torchsummary==1.5.1
torchvision==0.17.2+cu121
torchviz==0.0.2
tqdm==4.66.1
Utils==1.0.2
~orch==2.2.2
